# docs/contracts/http/schemaapi.openapi.yaml
openapi: 3.0.3
info:
  title: SchemaAPI
  version: 0.5.0
  description: Document Intelligence API with ingestion, hybrid retrieval, cited RAG, governance, agents, analysis reports and observability.
servers:
  - url: http://localhost:8081
paths:
...
  /governance/pii/redact:
    post:
      summary: Redact PII locally
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
      responses:
        "200":
          description: Redacted text and findings
...
  /governance/audit:
    get:
      summary: List recent audit events
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Audit events
  /observability/rag/evaluate:
    post:
      summary: Evaluate the latest audited RAG query
      responses:
        "200":
          description: Evaluation run
        "404":
          description: No query audit found
  /observability/rag/latest:
    get:
      summary: Get the latest RAG evaluation
      responses:
        "200":
          description: Latest evaluation run
        "404":
          description: No evaluation found
  /observability/rag/history:
    get:
      summary: List recent RAG evaluations
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: RAG evaluation history with quality metrics
...